remove feature flag, refactor to TS, and memo data
